Add nlohman::json library and use json as message data format
For simplicity, JSON is used as message data format. In the simplest
case, any messages can be used as an acknowledgement between services so
that the pipeline can work in order. However, we probably need to have
frame metadata not hard coded (as it is now).
Actual pixel data is sent using shared memory locations but frame
metadata info is sent using POSIX queues.